earthquake_api/requirements.txt
fastapi
uvicorn
python-dotenv 
praw
confluent-kafka